query builder

+++ src/models/Model.ts
@@ -1,34 +1,41 @@
-import _ from 'lodash';
-import { MODEL_FILTER_PATH } from '../config/app';
+import _ from "lodash";
+import { MODEL_FILTER_PATH } from "../config/app";
+import moment from 'moment'
+import { DATE_TIME_FORMAT } from "../lib/builder/config/sql";
 
 interface ModelGetInterface {
-  fields?: string[],
-  filters?: Record<string, any>,
-  limit?: number,
-  offset?: number
+  fields?: string[];
+  filters?: Record<string, any>;
+  limit?: number;
+  offset?: number;
 }
 
 interface ModelCountInterface {
-  filters?: Record<string, any>
+  filters?: Record<string, any>;
 }
 
 class Model {
   _tableName?: string;
-  _primary: string = 'id';
+  _primary: string = "id";
   _builder: any;
   _modelFilter: any;
-  #baseBuilder = require('../lib/builder')
+  _softDelete: string | false = "deleted_at";
+  _created_at: string | false = 'created_at';
+  _updated_at: string | false = 'updated_at';
+  #baseBuilder = require("../lib/builder");
 
   _init() {
-    this._builder = this.#baseBuilder.table(this._tableName || this)
+    this._builder = this.#baseBuilder.table(this._tableName || this);
 
-    if(!this._modelFilter) {
-      const className = `${this.constructor.name}Filter`
+    if (!this._modelFilter) {
+      const className = `${this.constructor.name}Filter`;
 
       try {
-        this._modelFilter = new (require(`${MODEL_FILTER_PATH}/${className}`))(this._builder);
+        this._modelFilter = new (require(`${MODEL_FILTER_PATH}/${className}`))(
+          this._builder
+        );
       } catch (e) {
-        console.warn('e :>> ', e);
+        console.warn("e :>> ", e);
       }
     }
   }
@@ -38,18 +45,35 @@ class Model {
       this.filters(filters);
     }
 
-    return this._builder.count(['*']);
+    return this._builder.count(["*"]);
   }
 
   find(id: number) {
+    if (this._softDelete) {
+      this._builder.whereNull(this._softDelete);
+    }
+
     return this._builder.where(this._primary, id).get();
   }
 
-  get({ fields, filters, limit, offset }: ModelGetInterface = {} as ModelGetInterface) {
-    this._builder.select(fields || ['*']);
+  get(
+    {
+      fields,
+      filters,
+      limit,
+      offset,
+    }: ModelGetInterface = {} as ModelGetInterface,
+    force: boolean = false
+  ) {
+    // const query = this.newQuery();
+    this._builder.select(fields || ["*"]);
 
     if (!_.isEmpty(filters)) {
       this.filters(filters);
+    }
+
+    if (this._softDelete && !force) {
+      this._builder.whereNull(this._softDelete);
     }
 
     if (limit) {
@@ -63,22 +87,39 @@ class Model {
     return this._builder.get();
   }
 
-  insert(values: Record<string, number|string> | Record<string, number|string>[]) {
-    return this._builder.insert(values);
+  insert(
+    values: Record<string, number | string> | Record<string, number | string>[]
+  ) {
+    if (!this._created_at && !this._updated_at) {
+      return this._builder.insert(values);
+    }
+
+    return this._builder.insert(this.#prepareInsertData(values));
   }
 
-  update(attribute: string | number, values: Record<string, string|number>) {
-    return this._builder.where(this._primary, attribute).update(values);
+  update(attribute: string | number, values: Record<string, string | number>) {
+    return this._builder.where(this._primary, attribute).update({
+      ...values,
+      ...(this._updated_at && { [this._updated_at]: moment().format(DATE_TIME_FORMAT) })
+    });
   }
 
-  delete(id: string | number) {
-    return this._builder.where(this._primary, id).delete();
+  delete(id: string | number, force: boolean = false) {
+    if (force || !this._softDelete) {
+      return this._builder.where(this._primary, id).delete();
+    }
+
+    return this._builder
+      .where(this._primary, id)
+      .update({
+        [this._softDelete]: moment().format(DATE_TIME_FORMAT),
+      });
   }
 
   filters(filters: Record<string, any>) {
     for (const field in filters) {
       if (Object.prototype.hasOwnProperty.call(filters, field)) {
-        this.#addFilter(field, filters[field])
+        this.#addFilter(field, filters[field]);
       }
     }
 
@@ -92,7 +133,7 @@ class Model {
   #addFilter(field: string, value: any) {
     const method = _.camelCase(field);
 
-    if(this._modelFilter && this._modelFilter[method]) {
+    if (this._modelFilter && this._modelFilter[method]) {
       return this._modelFilter[method].bind(this._builder)(value);
     }
 
@@ -102,6 +143,23 @@ class Model {
 
     return this._builder.where(field, value);
   }
+
+  #prepareInsertData(values: Record<string, number | string> | Record<string, number | string>[]) {
+    const timestamp = moment().format(DATE_TIME_FORMAT);
+    if (!Array.isArray(values)) {
+      return {
+        ...values,
+        ...(this._updated_at && { [this._updated_at]: timestamp }),
+        ...(this._created_at && { [this._created_at]: timestamp }),
+      };
+    }
+
+    return values.map(v => ({
+      ...v,
+      ...(this._updated_at && { [this._updated_at]: timestamp }),
+      ...(this._created_at && { [this._created_at]: timestamp }),
+    }))
+  }
 }
 
-module.exports = Model
+module.exports = Model;
